Save docket metadata in directory.

save_dockets("[agency_acronym]") creates a data directory for collecting agency docket metadata within a folder structure.

Usage

save_dockets <- function(agency)

Arguments

agency

Agency acronym(s) (see official acronyms on regulations.gov)

Details

This function uses the get_dockets function internally, so you will need to have an 'api_keys' stored in your environment in order for it to function properly.

Value

Stores data as a .rmd file.

References

Author

Note

See also

Examples

agencies <- c("NLRB", "OMB", "TSA")

save_dockets(agencies)
#> Error in save_dockets(agencies): could not find function "save_dockets"